Move tests to table visualization scenarios tests suite
