+++ e2e/test/scenarios/visualizations/table.cy.spec.js
@@ -13,6 +13,7 @@ describe("scenarios > visualizations > table", () => {
   beforeEach(() => {
     restore();
     cy.signInAsNormalUser();
+    cy.intercept("GET", "/api/field/*/search/*").as("findSuggestions");
   });
 
   function joinTable(table) {
@@ -268,6 +269,46 @@ describe("scenarios > visualizations > table", () => {
     // Cypress is too fast and is doing the assertions in that split second while popover is reloading which results in a false positive result.
     cy.wait(100);
     popover().should("not.exist");
+  });
+
+  it("popover should not be horizontally scrollable (metabase#31339)", () => {
+    openPeopleTable();
+    headerCells().filter(":contains('Password')").click();
+
+    popover().within(() => {
+      cy.findByText("Filter by this column").click();
+      cy.findByPlaceholderText("Search by Password").type("e").blur();
+      cy.wait("@findSuggestions");
+    });
+
+    popover().then(popoverElement => {
+      expect(
+        popoverElement[0].clientHeight,
+        "horizontal scrollbar is not shown",
+      ).to.eq(popoverElement[0].offsetHeight);
+    });
+  });
+
+  it("default picker container should not be horizontally scrollable", () => {
+    openPeopleTable();
+    headerCells().filter(":contains('Password')").click();
+
+    popover().within(() => {
+      cy.findByText("Filter by this column").click();
+
+      const input = cy.findByPlaceholderText("Search by Password");
+      input.type("e").blur();
+      cy.wait("@findSuggestions");
+      input.type("f");
+      cy.wait("@findSuggestions");
+
+      cy.findByTestId("default-picker-container").then(containerElement => {
+        expect(
+          containerElement[0].clientHeight,
+          "horizontal scrollbar is not shown",
+        ).to.eq(containerElement[0].offsetHeight);
+      });
+    });
   });
 });
